+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 ### Version 1.5.0:
 - Added Url support to event tooltips (can be clicked, which opens the Url in a new window/tab).
+- Only valid Urls can now be entered in the Event Editor dialog (an error message is shown for invalid entries).
+- Events that contain an invalid Url when added/updated via public functions are now cleared.
 - 
 
 ### Version 1.4.5: